feat(arch): split fat detail views into sub-components
  - PurchaseOrderDetailView (844 LOC) → 8 files, max 242 LOC each
  - AdjustmentDetailView (706 LOC) → 8 files, max 231 LOC each
  - Extracted mutations + dialog state into usePurchaseOrderActions /
    useAdjustmentActions hooks (SRP)
  - Created section components: Header, Info, Items, Receipts,
    EditDialog, SendDialog / ConfirmDialog
  - Replaced inline formatCurrency and toLocaleString calls with
    formatDate / formatDatetime / formatCurrency from @shared/lib/format
