+++ src/views/inventory/AdjustmentDetailView/EditAdjustmentDialog.tsx
@@ -1,0 +1,65 @@
+import Button from '@/components/ui/Button'
+import Dialog from '@/components/ui/Dialog'
+import Input from '@/components/ui/Input'
+import type { AdjustmentUpdateInput } from '@/services/AdjustmentService'
+
+type Props = {
+    open: boolean
+    editData: AdjustmentUpdateInput
+    isPending: boolean
+    onClose: () => void
+    onChange: (data: AdjustmentUpdateInput) => void
+    onSave: () => void
+}
+
+export const EditAdjustmentDialog = ({
+    open,
+    editData,
+    isPending,
+    onClose,
+    onChange,
+    onSave,
+}: Props) => (
+    <Dialog
+        isOpen={open}
+        width={500}
+        onClose={onClose}
+        onRequestClose={onClose}
+    >
+        <h5 className="mb-4">Editar Ajuste</h5>
+        <div className="space-y-4">
+            <div>
+                <label className="block text-sm font-medium mb-2">
+                    Responsable
+                </label>
+                <Input
+                    type="text"
+                    placeholder="Nombre del responsable"
+                    value={editData.adjustedBy || ''}
+                    onChange={(e) =>
+                        onChange({ ...editData, adjustedBy: e.target.value })
+                    }
+                />
+            </div>
+            <div>
+                <label className="block text-sm font-medium mb-2">Notas</label>
+                <Input
+                    textArea
+                    placeholder="Observaciones"
+                    value={editData.notes || ''}
+                    onChange={(e) =>
+                        onChange({ ...editData, notes: e.target.value })
+                    }
+                />
+            </div>
+        </div>
+        <div className="flex justify-end gap-2 mt-6">
+            <Button variant="plain" onClick={onClose}>
+                Cancelar
+            </Button>
+            <Button variant="solid" loading={isPending} onClick={onSave}>
+                Guardar
+            </Button>
+        </div>
+    </Dialog>
+)
